fix: update gsm8k to new math_benchmark base
